+++ SpawnDev.SpawnJS/JSObjects/DOM/DOMParser.cs
@@ -30,6 +30,6 @@ namespace SpawnDev.SpawnJS.JSObjects
         /// - application/xhtml+xml<br/>
         /// - image/svg+xml</param>
         /// <returns></returns>
-        public Document ParseFromString(string input, string mimeType) =>JSRef!.Call<Document>(input, mimeType);
+        public Document ParseFromString(string input, string mimeType) =>JSRef!.Call<Document>("parseFromString", input, mimeType);
     }
 }